ALteração Tipo dado

+++ src/main/java/com/senai/atvd/Entity/Venda.java
@@ -19,7 +19,7 @@ public class Venda {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idVenda;
+    private Long idVenda;
 
     @OneToMany(mappedBy = "venda", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Produto> produto;
